feat: 엔티티 클래스에 @Column 애너테이션 추가
1. 엔티티 클래스에 @Column 애너테이션을 추가하여 칼 속성을 지정한다
2. 엔티티의 공통 부분을 BaseEntity 클래스로 추출하여 이를 상속하도록 수정한다
3. BaseEntity 클래스에 생성 날짜 및 최근 수정 날짜 필드를 생성한다
