+++ src/main/java/techcourse/myblog/MyblogApplication.java
@@ -2,7 +2,9 @@ package techcourse.myblog;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
+import org.springframework.data.jpa.repository.config.EnableJpaAuditing;
 
+@EnableJpaAuditing
 @SpringBootApplication
 public class MyblogApplication {
