+++ src/test/java/techcourse/myblog/MyblogApplicationTest.java
@@ -3,9 +3,9 @@ package techcourse.myblog;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
 import org.springframework.boot.test.context.SpringBootTest;
+import org.springframework.data.jpa.repository.config.EnableJpaAuditing;
 import org.springframework.test.context.junit.jupiter.SpringExtension;
 
-@ExtendWith(SpringExtension.class)
 @SpringBootTest
 public class MyblogApplicationTest {
     @Test
